refactor unit tests
